+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - Added option to show/hide current weather text
+- Added option to set day format in forecast table
 
 ### Fixed
 - Added icon mapping for RBB icon 333001 (thunder)